+++ app/frontend/biz/service/get_cart.go
@@ -25,7 +25,7 @@ func NewGetCartService(Context context.Context, RequestContext *app.RequestConte
 
 func (h *GetCartService) Run(req *common.Empty) (resp map[string]any, err error) {
 	rpcResp, err := rpc.CartClient.GetCart(h.Context, &rpccart.GetCartReq{
-		UserId: uint32(frontendutils.GetUserIdFormCtx(h.Context)),
+		UserId: frontendutils.GetUserIdFormCtx(h.Context),
 	})
 	if err != nil {
 		return nil, err
