+++ app/product/biz/service/get_product.go
@@ -31,7 +31,7 @@ func (s *GetProductService) Run(req *product.GetProductReq) (resp *product.GetPr
 
 	return &product.GetProductResp{
 		Product: &product.Product{
-			Id:          uint32(p.ID),
+			Id:          p.ID,
 			Picture:     p.Picture,
 			Price:       p.Price,
 			Description: p.Description,
